LayoutTests/imported/w3c/web-platform-tests/offscreen-canvas/drawing-rectangles-to-the-canvas/2d.clearRect.globalcomposite-expected.txt

2d.clearRect.globalcomposite

clearRect is not affected by globalCompositeOperation


PASS clearRect is not affected by globalCompositeOperation 
